Small Fixes
Fixed an issue where the "Scripts Forum" could disappear from the Wazebar when changes were made to settings. The problem occurred because there were two <div> elements with the same ID in the DOM, which could result in it being incorrectly located.

Updated the logic in "CheckForum" to only fetch URLs containing /c, /search, or /tag, as these are the only options with topics. This change removes the console warning that appeared for new WIKI URLs when added to the Custom Link section.

Bumped Version number.